requirement.txt
streamlit
tensorflow
numpy
pandas
opencv-python-headless
Pillow
scikit-learn